Temporarily, don't use structured ARN (details below)
aws-general provides a nice structured Arn type, but sadly, it uses a
fixed enumerated type called ServiceNamespace; every time Amazon
releases a new service, then this type in aws-general must be updated,
which is pretty unfortunate.

Until that type either gets updated or is generalized to be modular with
respect to services, I'll just use plain text as the type for ARNs.